x86_64 compile fixes

+++ src/cfio/control.cpp
@@ -440,8 +440,8 @@ Handle Control::Alloc(size_t Size, uint Flag, const std::string& Owner,
 // GlobalRealloc
 Handle Control::ReAlloc(Handle hMemory, uint32_t wNewSize, uint32_t wFlag) {
 	Handle hNewMemory;
-	uint32_t wOldSize = 0;
-	uint32_t wOldFlag = 0;
+	size_t wOldSize = 0;
+	uint wOldFlag = 0;
 	uint32_t GlobalFlag = 0;
 	void * Sorc;
 	void * Desc;
@@ -505,7 +505,7 @@ bool Control::Unlock(Handle hMem) {
 
 uint32_t Control::WriteMemToFile(Handle hMem, const std::string& Name) {
 	Handle hFile = OpenFile(NULL, Name, OSF_WRITE);
-	uint32_t wMemorySize;
+	size_t wMemorySize;
 	uint32_t wMemoryFlag;
 	uint32_t Counter = 0;
 	pchar pMem;
@@ -624,7 +624,7 @@ bool Control::AddNewMemoryInList(Handle hMemory, uint Size, uint32_t IsGlobal,
 	return memory_list_.AddItem(hMemory, Size, IsGlobal, Owner, Comment);
 }
 
-bool Control::TakeMemory(Handle hMemory, uint32_t * wMemorySize,
+bool Control::TakeMemory(Handle hMemory, size_t * wMemorySize,
 		uint32_t * wMemoryFlag) {
 	return memory_list_.TakeItem(hMemory, wMemorySize, wMemoryFlag);
 }
@@ -644,7 +644,7 @@ bool Control::LockatorMemoryInList(Handle hMemory, Bool32 bLock) {
 
 bool Control::FreeMemory(Handle hMemory, uint32_t /*wFlag*/) {
 	uint32_t wMemoryStatus;
-	uint32_t wMemorySize;
+	size_t wMemorySize;
 
 	if (TakeMemory(hMemory, &wMemorySize, &wMemoryStatus)) {
 		// delete header from list
@@ -659,7 +659,7 @@ bool Control::FreeMemory(Handle hMemory, uint32_t /*wFlag*/) {
 
 void * Control::LockMemory(Handle hMemory) {
 	uint32_t wMemoryStatus;
-	uint32_t wMemorySize;
+	size_t wMemorySize;
 	pvoid pMemory = NULL;
 	// ну , тут осталось немного. отписывать адрес в память,
 	// но пока тут падает, а так как глобальная память не используется, то и
@@ -685,7 +685,7 @@ void * Control::LockMemory(Handle hMemory) {
 
 bool Control::UnlockMemory(Handle hMemory) {
 	uint32_t wMemoryStatus;
-	uint32_t wMemorySize;
+	size_t wMemorySize;
 	Bool bUnlock = false;
 
 	if (TakeMemory(hMemory, &wMemorySize, &wMemoryStatus)) {
