+++ src/rstr/rstr_p2.cpp
@@ -409,8 +409,8 @@ int32_t p2_Cstr2Cell(CSTR_line lin, CSTR_rast first, CSTR_rast last,
 		c1 = c2;
 
 		// Nick add
-		c2->dupstart = (int32_t) CSTR_GetDup(curr);
-		c2->dupend = (int32_t) CSTR_GetDupEnd(curr);
+		c2->dupstart = CSTR_GetDup(curr);
+		c2->dupend = CSTR_GetDupEnd(curr);
 
 		numCell++;
 	}
